Add Admin.WebUI & Update IdentityServer.WebUI
- Add Eiromplays.IdentityServer.Admin.WebUI which will be used for the admin dashboard

- Update  Eiromplays.IdentityServer.WebUI to use react-toastify instead of it's own notification system, also added roles to the session info, plans for the session info is to get everything from it, instead of fetching from the users endpoint and the session info endpoint

- Update API, fixed the users endpoint, it should now work, and you need Administrator role to access it

+++ src/Application/Eiromplays.IdentityServer.Application/DTOs/User/UserDto.cs
@@ -28,7 +28,7 @@ public class UserDto
 
     public DateTimeOffset? LockoutEnd { get; set; }
 
-    public decimal Credits { get; set; }
+    public double Credits { get; set; }
 
     public string? DiscordId { get; set; }
 }
